requirements.txt
fastapi
passkeys>=1.1.3
Jinja2
uvicorn
python-dotenv